Revert "Fixing onClick priority"
This reverts commit 5892b1f82d673f2ebd1a1489a93e1d97fead5676.
