+++ app/src/main/java/com/jerboa/MainActivity.kt
@@ -11,10 +11,9 @@ import android.widget.Toast
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideOutHorizontally
-import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
@@ -95,6 +94,7 @@ class MainActivity : AppCompatActivity() {
         AppSettingsViewModelFactory((application as JerboaApplication).appSettingsRepository)
     }
 
+    @OptIn(ExperimentalAnimationApi::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -148,8 +148,6 @@ class MainActivity : AppCompatActivity() {
                     else -> {}
                 }
 
-                val drawerState = rememberDrawerState(DrawerValue.Closed)
-
                 NavHost(
                     route = Route.Graph.ROOT,
                     navController = navController,
@@ -179,7 +177,6 @@ class MainActivity : AppCompatActivity() {
                             siteViewModel = siteViewModel,
                             appSettingsViewModel = appSettingsViewModel,
                             appSettings = appSettings,
-                            drawerState = drawerState,
                         )
                     }
 
@@ -309,7 +306,6 @@ class MainActivity : AppCompatActivity() {
                             useCustomTabs = appSettings.useCustomTabs,
                             usePrivateTabs = appSettings.usePrivateTabs,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
 
@@ -341,7 +337,6 @@ class MainActivity : AppCompatActivity() {
                             useCustomTabs = appSettings.useCustomTabs,
                             usePrivateTabs = appSettings.usePrivateTabs,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
 
@@ -361,7 +356,6 @@ class MainActivity : AppCompatActivity() {
                             siteViewModel = siteViewModel,
                             selectMode = args.select,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
 
@@ -414,7 +408,6 @@ class MainActivity : AppCompatActivity() {
                             accountViewModel = accountViewModel,
                             siteViewModel = siteViewModel,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
