+++ app/src/main/java/com/jerboa/model/InboxViewModel.kt
@@ -2,7 +2,6 @@ package com.jerboa.model
 
 import android.content.Context
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
@@ -79,11 +78,11 @@ class InboxViewModel : ViewModel(), Initializable {
     private var blockPersonRes: ApiState<BlockPersonResponse> by
         mutableStateOf(ApiState.Empty)
 
-    var pageReplies by mutableIntStateOf(1)
+    var pageReplies by mutableStateOf(1)
         private set
-    var pageMentions by mutableIntStateOf(1)
+    var pageMentions by mutableStateOf(1)
         private set
-    var pageMessages by mutableIntStateOf(1)
+    var pageMessages by mutableStateOf(1)
         private set
     var unreadOnly by mutableStateOf(true)
         private set
@@ -99,9 +98,9 @@ class InboxViewModel : ViewModel(), Initializable {
     }
 
     fun resetPages() {
-        resetPageMentions()
-        resetPageMessages()
-        resetPageReplies()
+        pageMentions = 1
+        pageReplies = 1
+        pageMessages = 1
     }
 
     fun updateUnreadOnly(unreadOnly: Boolean) {
@@ -515,23 +514,6 @@ class InboxViewModel : ViewModel(), Initializable {
             unread_only = unreadOnly,
             sort = CommentSortType.New,
             page = pageReplies,
-            auth = jwt,
-        )
-    }
-
-    fun getFormMentions(jwt: String): GetPersonMentions {
-        return GetPersonMentions(
-            unread_only = unreadOnly,
-            sort = CommentSortType.New,
-            page = pageMentions,
-            auth = jwt,
-        )
-    }
-
-    fun getFormMessages(jwt: String): GetPrivateMessages {
-        return GetPrivateMessages(
-            unread_only = unreadOnly,
-            page = pageMessages,
             auth = jwt,
         )
     }
