+++ app/src/main/java/com/jerboa/ui/components/common/InputFields.kt
@@ -59,6 +59,7 @@ import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.input.getSelectedText
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import com.jerboa.R
 import com.jerboa.api.uploadPictrsImage
@@ -580,9 +581,22 @@ fun TextMarkdownBarPreview() {
 }
 
 @Composable
+fun PreviewLines(
+    text: String,
+    modifier: Modifier = Modifier,
+) {
+    Text(
+        text = text,
+        style = MaterialTheme.typography.bodyMedium,
+        maxLines = 5,
+        overflow = TextOverflow.Ellipsis,
+        modifier = modifier,
+    )
+}
+
+@Composable
 fun MyMarkdownText(
     markdown: String,
-    modifier: Modifier = Modifier,
     color: Color = MaterialTheme.colorScheme.onSurface,
     onClick: () -> Unit,
     onLongClick: (() -> Unit)? = null,
@@ -592,7 +606,6 @@ fun MyMarkdownText(
         color = color,
         onClick = onClick,
         onLongClick = onLongClick,
-        modifier = modifier,
     )
 }
