+++ app/src/main/java/com/jerboa/ui/components/community/list/CommunityList.kt
@@ -21,9 +21,12 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
 import com.jerboa.R
 import com.jerboa.datatypes.sampleCommunityView
 import com.jerboa.datatypes.types.*
+import com.jerboa.ui.components.common.DefaultBackButton
 import com.jerboa.ui.components.common.simpleVerticalScrollbar
 import com.jerboa.ui.components.community.CommunityLinkLarger
 import com.jerboa.ui.components.community.CommunityLinkLargerWithUserCount
@@ -31,7 +34,7 @@ import com.jerboa.ui.components.community.CommunityLinkLargerWithUserCount
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun CommunityListHeader(
-    openDrawer: () -> Unit,
+    navController: NavController = rememberNavController(),
     search: String,
     onSearchChange: (search: String) -> Unit,
 ) {
@@ -53,14 +56,7 @@ fun CommunityListHeader(
                 )
             }
         },
-        navigationIcon = {
-            IconButton(onClick = openDrawer) {
-                Icon(
-                    Icons.Outlined.Menu,
-                    contentDescription = stringResource(R.string.home_menu),
-                )
-            }
-        },
+        navigationIcon = { DefaultBackButton(navController) },
     )
 }
