+++ app/src/main/java/com/jerboa/ui/components/home/BottomNavActivity.kt
@@ -1,24 +1,15 @@
 package com.jerboa.ui.components.home
 
+import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Bookmarks
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.outlined.Bookmarks
-import androidx.compose.material.icons.outlined.Email
-import androidx.compose.material.icons.outlined.Home
-import androidx.compose.material.icons.outlined.Person
-import androidx.compose.material.icons.outlined.Search
-import androidx.compose.material3.DrawerState
+import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ModalDrawerSheet
 import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -27,7 +18,6 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTagsAsResourceId
@@ -38,7 +28,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import arrow.core.Either
-import com.jerboa.R
 import com.jerboa.api.ApiState
 import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettings
@@ -51,27 +40,17 @@ import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.InitializeRoute
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.community.list.CommunityListActivity
-import com.jerboa.ui.components.drawer.MainDrawer
 import com.jerboa.ui.components.inbox.InboxActivity
 import com.jerboa.ui.components.person.PersonProfileActivity
 
-enum class NavTab(
-    val textId: Int,
-    val iconOutlined: ImageVector,
-    val iconFilled: ImageVector,
-    val contentDescriptionId: Int,
-) {
-    Home(R.string.bottomBar_label_home, Icons.Outlined.Home, Icons.Filled.Home, R.string.bottomBar_home),
-    Search(R.string.bottomBar_label_search, Icons.Outlined.Search, Icons.Filled.Search, R.string.bottomBar_search),
-    Inbox(R.string.bottomBar_label_inbox, Icons.Outlined.Email, Icons.Filled.Email, R.string.bottomBar_inbox),
-    Saved(R.string.bottomBar_label_bookmarks, Icons.Outlined.Bookmarks, Icons.Filled.Bookmarks, R.string.bottomBar_bookmarks),
-    Profile(R.string.bottomBar_label_profile, Icons.Outlined.Person, Icons.Filled.Person, R.string.bottomBar_profile),
-    ;
+enum class BottomNavTab {
+    Home, Search, Inbox, Saved, Profile;
 
     fun needsLogin() = this == Inbox || this == Saved || this == Profile
 }
 
 @OptIn(
+    ExperimentalAnimationApi::class,
     ExperimentalLayoutApi::class,
     ExperimentalComposeUiApi::class,
 )
@@ -82,25 +61,26 @@ fun BottomNavActivity(
     siteViewModel: SiteViewModel,
     appSettingsViewModel: AppSettingsViewModel,
     appSettings: AppSettings,
-    drawerState: DrawerState,
 ) {
     val account = getCurrentAccount(accountViewModel)
     val ctx = LocalContext.current
     val scope = rememberCoroutineScope()
 
     val bottomNavController = rememberNavController()
-    var selectedTab by rememberSaveable { mutableStateOf(NavTab.Home) }
-    val onSelectTab = { tab: NavTab ->
+    var selectedTab by rememberSaveable { mutableStateOf(BottomNavTab.Home) }
+    val onSelectTab = { tab: BottomNavTab ->
         if (tab.needsLogin() && account == null) {
             loginFirstToast(ctx)
         } else {
             selectedTab = tab
             bottomNavController.navigate(tab.name) {
                 launchSingleTop = true
-                popUpTo(bottomNavController.graph.id) // To make back button close the app.
+                popUpTo(0) // To make back button close the app.
             }
         }
     }
+
+    val drawerState = rememberDrawerState(DrawerValue.Closed)
 
     val homeViewModel: HomeViewModel = viewModel()
     if (siteViewModel.siteRes is ApiState.Success) {
@@ -112,7 +92,7 @@ fun BottomNavActivity(
     }
 
     ModalNavigationDrawer(
-        gesturesEnabled = true,
+        gesturesEnabled = selectedTab == BottomNavTab.Home,
         drawerState = drawerState,
         drawerContent = {
             ModalDrawerSheet(
@@ -124,9 +104,8 @@ fun BottomNavActivity(
                         homeViewModel = homeViewModel,
                         scope = scope,
                         drawerState = drawerState,
-                        onSelectTab = onSelectTab,
+                        onSelectTab = if (appSettings.showBottomNav) onSelectTab else null,
                         blurNSFW = appSettings.blurNSFW,
-                        showBottomNav = appSettings.showBottomNav,
                     )
                 },
             )
@@ -135,18 +114,17 @@ fun BottomNavActivity(
         content = {
             Scaffold(
                 bottomBar = {
-                    if (appSettings.showBottomNav) {
-                        BottomAppBarAll(
-                            selectedTab = selectedTab,
-                            unreadCounts = siteViewModel.getUnreadCountTotal(),
-                            showTextDescriptionsInNavbar = appSettings.showTextDescriptionsInNavbar,
-                            onSelect = onSelectTab,
-                        )
-                    }
+                    BottomAppBarAll(
+                        showBottomNav = appSettings.showBottomNav,
+                        selectedTab = selectedTab,
+                        unreadCounts = siteViewModel.getUnreadCountTotal(),
+                        showTextDescriptionsInNavbar = appSettings.showTextDescriptionsInNavbar,
+                        onSelect = onSelectTab,
+                    )
                 },
             ) { padding ->
                 val bottomPadding =
-                    if (selectedTab == NavTab.Search && WindowInsets.isImeVisible) {
+                    if (selectedTab == BottomNavTab.Search && WindowInsets.isImeVisible) {
                         0.dp
                     } else {
                         padding.calculateBottomPadding()
@@ -154,10 +132,10 @@ fun BottomNavActivity(
 
                 NavHost(
                     navController = bottomNavController,
-                    startDestination = NavTab.Home.name,
+                    startDestination = BottomNavTab.Home.name,
                     modifier = Modifier.padding(bottom = bottomPadding),
                 ) {
-                    composable(route = NavTab.Home.name) {
+                    composable(route = BottomNavTab.Home.name) {
                         HomeActivity(
                             navController = navController,
                             homeViewModel = homeViewModel,
@@ -172,28 +150,26 @@ fun BottomNavActivity(
                         )
                     }
 
-                    composable(route = NavTab.Search.name) {
+                    composable(route = BottomNavTab.Search.name) {
                         CommunityListActivity(
                             navController = navController,
                             accountViewModel = accountViewModel,
                             selectMode = false,
                             siteViewModel = siteViewModel,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
 
-                    composable(route = NavTab.Inbox.name) {
+                    composable(route = BottomNavTab.Inbox.name) {
                         InboxActivity(
                             navController = navController,
                             accountViewModel = accountViewModel,
                             siteViewModel = siteViewModel,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
 
-                    composable(route = NavTab.Saved.name) {
+                    composable(route = BottomNavTab.Saved.name) {
                         PersonProfileActivity(
                             personArg = Either.Left(account!!.id),
                             savedMode = true,
@@ -205,11 +181,10 @@ fun BottomNavActivity(
                             useCustomTabs = appSettings.useCustomTabs,
                             usePrivateTabs = appSettings.usePrivateTabs,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
 
-                    composable(route = NavTab.Profile.name) {
+                    composable(route = BottomNavTab.Profile.name) {
                         PersonProfileActivity(
                             personArg = Either.Left(account!!.id),
                             savedMode = false,
@@ -221,7 +196,6 @@ fun BottomNavActivity(
                             useCustomTabs = appSettings.useCustomTabs,
                             usePrivateTabs = appSettings.usePrivateTabs,
                             blurNSFW = appSettings.blurNSFW,
-                            drawerState = drawerState,
                         )
                     }
                 }
