+++ app/src/main/java/com/jerboa/ui/components/home/Home.kt
@@ -1,16 +1,43 @@
 package com.jerboa.ui.components.home
 
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Add
+import androidx.compose.material.icons.outlined.Bookmarks
+import androidx.compose.material.icons.outlined.Close
+import androidx.compose.material.icons.outlined.Email
+import androidx.compose.material.icons.outlined.ExpandLess
+import androidx.compose.material.icons.outlined.ExpandMore
 import androidx.compose.material.icons.outlined.FilterList
 import androidx.compose.material.icons.outlined.Info
+import androidx.compose.material.icons.outlined.List
+import androidx.compose.material.icons.outlined.LocationCity
+import androidx.compose.material.icons.outlined.Login
 import androidx.compose.material.icons.outlined.Menu
 import androidx.compose.material.icons.outlined.MoreVert
+import androidx.compose.material.icons.outlined.Person
+import androidx.compose.material.icons.outlined.Public
 import androidx.compose.material.icons.outlined.Refresh
+import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material.icons.outlined.Sort
 import androidx.compose.material.icons.outlined.ViewAgenda
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Divider
+import androidx.compose.material3.DrawerState
+import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -19,11 +46,15 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.TopAppBarScrollBehavior
+import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -31,24 +62,415 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTagsAsResourceId
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.jerboa.PostViewMode
 import com.jerboa.R
+import com.jerboa.api.ApiState
+import com.jerboa.datatypes.samplePerson
+import com.jerboa.datatypes.types.Community
+import com.jerboa.datatypes.types.GetSiteResponse
 import com.jerboa.datatypes.types.ListingType
+import com.jerboa.datatypes.types.MyUserInfo
+import com.jerboa.datatypes.types.Person
 import com.jerboa.datatypes.types.SortType
 import com.jerboa.datatypes.types.Tagline
+import com.jerboa.db.Account
+import com.jerboa.db.AccountViewModel
+import com.jerboa.federatedNameShown
 import com.jerboa.getLocalizedListingTypeName
 import com.jerboa.getLocalizedSortingTypeShortName
 import com.jerboa.ui.components.common.IconAndTextDrawerItem
+import com.jerboa.ui.components.common.LargerCircularIcon
 import com.jerboa.ui.components.common.ListingTypeOptionsDialog
 import com.jerboa.ui.components.common.MyMarkdownText
+import com.jerboa.ui.components.common.PictrsBannerImage
 import com.jerboa.ui.components.common.PostViewModeDialog
 import com.jerboa.ui.components.common.SortOptionsDialog
 import com.jerboa.ui.components.common.SortTopOptionsDialog
+import com.jerboa.ui.components.common.simpleVerticalScrollbar
+import com.jerboa.ui.components.common.toLogin
 import com.jerboa.ui.components.common.toSiteSideBar
+import com.jerboa.ui.components.community.CommunityLinkLarger
+import com.jerboa.ui.components.person.PersonName
+import com.jerboa.ui.theme.DRAWER_BANNER_SIZE
 import com.jerboa.ui.theme.LARGE_PADDING
+import com.jerboa.ui.theme.SMALL_PADDING
+import com.jerboa.ui.theme.XL_PADDING
+import com.jerboa.ui.theme.muted
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+
+@Composable
+fun Drawer(
+    siteRes: ApiState<GetSiteResponse>,
+    unreadCount: Int,
+    navController: NavController = rememberNavController(),
+    accountViewModel: AccountViewModel,
+    onSwitchAccountClick: (account: Account) -> Unit,
+    onSignOutClick: () -> Unit,
+    onClickListingType: (ListingType) -> Unit,
+    onCommunityClick: (community: Community) -> Unit,
+    onClickProfile: () -> Unit,
+    onClickInbox: () -> Unit,
+    onClickSaved: () -> Unit,
+    onClickSettings: () -> Unit,
+    onClickCommunities: () -> Unit,
+    isOpen: Boolean,
+    blurNSFW: Boolean,
+) {
+    var showAccountAddMode by rememberSaveable { mutableStateOf(false) }
+
+    if (!isOpen) showAccountAddMode = false
+
+    val myUserInfo = when (siteRes) {
+        is ApiState.Success -> siteRes.data.my_user
+        else -> null
+    }
+
+    DrawerHeader(
+        myPerson = myUserInfo?.local_user_view?.person,
+        showAccountAddMode = showAccountAddMode,
+        onClickShowAccountAddMode = { showAccountAddMode = !showAccountAddMode },
+        showAvatar = myUserInfo?.local_user_view?.local_user?.show_avatars ?: true,
+    )
+    // Drawer items
+    DrawerContent(
+        accountViewModel = accountViewModel,
+        unreadCount = unreadCount,
+        myUserInfo = myUserInfo,
+        showAccountAddMode = showAccountAddMode,
+        navController = navController,
+        onSwitchAccountClick = onSwitchAccountClick,
+        onSignOutClick = onSignOutClick,
+        onClickListingType = onClickListingType,
+        onCommunityClick = onCommunityClick,
+        onClickProfile = onClickProfile,
+        onClickInbox = onClickInbox,
+        onClickSaved = onClickSaved,
+        onClickSettings = onClickSettings,
+        onClickCommunities = onClickCommunities,
+        blurNSFW = blurNSFW,
+    )
+}
+
+@Composable
+fun DrawerContent(
+    showAccountAddMode: Boolean,
+    navController: NavController,
+    accountViewModel: AccountViewModel,
+    onSwitchAccountClick: (account: Account) -> Unit,
+    onSignOutClick: () -> Unit,
+    onClickListingType: (ListingType) -> Unit,
+    onCommunityClick: (community: Community) -> Unit,
+    onClickProfile: () -> Unit,
+    onClickInbox: () -> Unit,
+    onClickSaved: () -> Unit,
+    onClickSettings: () -> Unit,
+    onClickCommunities: () -> Unit,
+    myUserInfo: MyUserInfo?,
+    unreadCount: Int,
+    blurNSFW: Boolean,
+) {
+    AnimatedVisibility(
+        visible = showAccountAddMode,
+        enter = expandVertically(),
+        exit = shrinkVertically(),
+    ) {
+        Divider()
+        DrawerAddAccountMode(
+            accountViewModel = accountViewModel,
+            navController = navController,
+            onSwitchAccountClick = onSwitchAccountClick,
+            onSignOutClick = onSignOutClick,
+        )
+    }
+
+    Divider()
+    DrawerItemsMain(
+        myUserInfo = myUserInfo,
+        onClickListingType = onClickListingType,
+        onCommunityClick = onCommunityClick,
+        onClickProfile = onClickProfile,
+        onClickInbox = onClickInbox,
+        onClickSaved = onClickSaved,
+        unreadCount = unreadCount,
+        onClickSettings = onClickSettings,
+        onClickCommunities = onClickCommunities,
+        blurNSFW = blurNSFW,
+    )
+}
+
+@Composable
+fun DrawerItemsMain(
+    myUserInfo: MyUserInfo?,
+    onClickSaved: () -> Unit,
+    onClickProfile: () -> Unit,
+    onClickInbox: () -> Unit,
+    onClickSettings: () -> Unit,
+    onClickCommunities: () -> Unit,
+    onClickListingType: (ListingType) -> Unit,
+    onCommunityClick: (community: Community) -> Unit,
+    unreadCount: Int,
+    blurNSFW: Boolean,
+) {
+    val listState = rememberLazyListState()
+
+    val follows = myUserInfo?.follows
+
+    LazyColumn(
+        state = listState,
+        modifier = Modifier.simpleVerticalScrollbar(listState),
+    ) {
+        if (!follows.isNullOrEmpty()) {
+            item {
+                IconAndTextDrawerItem(
+                    text = stringResource(R.string.home_subscribed),
+                    icon = Icons.Outlined.Bookmarks,
+                    onClick = { onClickListingType(ListingType.Subscribed) },
+                )
+            }
+        }
+        item {
+            IconAndTextDrawerItem(
+                text = stringResource(R.string.home_local),
+                icon = Icons.Outlined.LocationCity,
+                onClick = { onClickListingType(ListingType.Local) },
+            )
+        }
+        item {
+            IconAndTextDrawerItem(
+                text = stringResource(R.string.home_all),
+                icon = Icons.Outlined.Public,
+                onClick = { onClickListingType(ListingType.All) },
+            )
+        }
+        item {
+            myUserInfo?.also {
+                IconAndTextDrawerItem(
+                    text = stringResource(R.string.home_saved),
+                    icon = Icons.Outlined.Bookmarks,
+                    onClick = onClickSaved,
+                )
+            }
+        }
+        item {
+            IconAndTextDrawerItem(
+                text = stringResource(R.string.home_communities),
+                icon = Icons.Outlined.List,
+                onClick = onClickCommunities,
+            )
+        }
+        item {
+            Divider()
+        }
+        item {
+            myUserInfo?.also {
+                IconAndTextDrawerItem(
+                    text = stringResource(R.string.home_profile),
+                    icon = Icons.Outlined.Person,
+                    onClick = onClickProfile,
+                )
+            }
+        }
+        item {
+            myUserInfo?.also {
+                IconAndTextDrawerItem(
+                    text = stringResource(R.string.home_inbox),
+                    icon = Icons.Outlined.Email,
+                    onClick = onClickInbox,
+                    iconBadgeCount = unreadCount,
+                )
+            }
+        }
+        item {
+            IconAndTextDrawerItem(
+                text = stringResource(R.string.home_settings),
+                icon = Icons.Outlined.Settings,
+                onClick = onClickSettings,
+            )
+        }
+        item {
+            myUserInfo?.also {
+                Divider()
+            }
+        }
+
+        follows?.also { follows ->
+            item {
+                Text(
+                    text = stringResource(R.string.home_subscriptions),
+                    modifier = Modifier.padding(LARGE_PADDING),
+                    color = MaterialTheme.colorScheme.onBackground.muted,
+                )
+            }
+            items(
+                follows,
+                key = { follow -> follow.community.id },
+            ) { follow ->
+                CommunityLinkLarger(
+                    community = follow.community,
+                    onClick = onCommunityClick,
+                    showDefaultIcon = true,
+                    blurNSFW = blurNSFW,
+                )
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+fun DrawerItemsMainPreview() {
+    DrawerItemsMain(
+        myUserInfo = null,
+        onClickListingType = {},
+        onClickProfile = {},
+        onClickInbox = {},
+        onCommunityClick = {},
+        onClickSaved = {},
+        onClickSettings = {},
+        onClickCommunities = {},
+        unreadCount = 2,
+        blurNSFW = true,
+    )
+}
+
+@Composable
+fun DrawerAddAccountMode(
+    navController: NavController = rememberNavController(),
+    accountViewModel: AccountViewModel?,
+    onSwitchAccountClick: (account: Account) -> Unit,
+    onSignOutClick: () -> Unit,
+) {
+    val accountsWithoutCurrent = accountViewModel?.allAccounts?.value?.toMutableList()
+    val currentAccount = accountsWithoutCurrent?.firstOrNull { it.current }
+    accountsWithoutCurrent?.remove(currentAccount)
+
+    Column {
+        IconAndTextDrawerItem(
+            text = stringResource(R.string.home_add_account),
+            icon = Icons.Outlined.Add,
+            onClick = { navController.toLogin() },
+        )
+        accountsWithoutCurrent?.forEach {
+            IconAndTextDrawerItem(
+                text = stringResource(R.string.home_switch_to, it.name, it.instance),
+                icon = Icons.Outlined.Login,
+                onClick = { onSwitchAccountClick(it) },
+            )
+        }
+        currentAccount?.also {
+            IconAndTextDrawerItem(
+                text = stringResource(R.string.home_sign_out),
+                icon = Icons.Outlined.Close,
+                onClick = onSignOutClick,
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun DrawerAddAccountModePreview() {
+    DrawerAddAccountMode(
+        onSignOutClick = {},
+        onSwitchAccountClick = {},
+        accountViewModel = null,
+    )
+}
+
+@Composable
+fun DrawerHeader(
+    myPerson: Person?,
+    onClickShowAccountAddMode: () -> Unit,
+    showAccountAddMode: Boolean = false,
+    showAvatar: Boolean,
+) {
+    val sizeMod = Modifier
+        .fillMaxWidth()
+        .height(DRAWER_BANNER_SIZE)
+
+    Box(
+        modifier = sizeMod
+            .clickable(onClick = onClickShowAccountAddMode),
+    ) {
+        myPerson?.banner?.also {
+            PictrsBannerImage(
+                url = it,
+            )
+        }
+        // banner
+        Row(
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = sizeMod
+                .padding(XL_PADDING),
+        ) {
+            Box(modifier = Modifier.weight(0.9f)) {
+                AvatarAndAccountName(myPerson, showAvatar)
+            }
+            Icon(
+                modifier = Modifier.weight(0.1f),
+                imageVector = if (showAccountAddMode) {
+                    Icons.Outlined.ExpandLess
+                } else {
+                    Icons.Outlined.ExpandMore
+                },
+                contentDescription = if (showAccountAddMode) {
+                    stringResource(R.string.moreOptions)
+                } else {
+                    stringResource(R.string.lessOptions)
+                },
+            )
+        }
+    }
+}
+
+@Composable
+fun AvatarAndAccountName(myPerson: Person?, showAvatar: Boolean) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(SMALL_PADDING),
+    ) {
+        if (showAvatar) {
+            myPerson?.avatar?.also {
+                LargerCircularIcon(icon = it)
+            }
+        }
+        Column() {
+            PersonName(
+                person = myPerson,
+                color = MaterialTheme.colorScheme.onSurface,
+            )
+            Text(
+                text = myPerson?.let { federatedNameShown(it) } ?: "",
+                color = MaterialTheme.colorScheme.tertiary,
+                style = MaterialTheme.typography.labelSmall,
+                overflow = TextOverflow.Ellipsis,
+                maxLines = 2,
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun AvatarAndAccountNamePreview() {
+    AvatarAndAccountName(myPerson = samplePerson, showAvatar = true)
+}
+
+@Preview
+@Composable
+fun DrawerHeaderPreview() {
+    DrawerHeader(
+        myPerson = samplePerson,
+        onClickShowAccountAddMode = {},
+        showAvatar = true,
+    )
+}
 
 @Composable
 fun HomeHeaderTitle(
@@ -71,7 +493,8 @@ fun HomeHeaderTitle(
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun HomeHeader(
-    openDrawer: () -> Unit,
+    scope: CoroutineScope,
+    drawerState: DrawerState,
     onClickSortType: (SortType) -> Unit,
     onClickListingType: (ListingType) -> Unit,
     onClickRefresh: () -> Unit,
@@ -156,7 +579,11 @@ fun HomeHeader(
             )
         },
         navigationIcon = {
-            IconButton(onClick = openDrawer) {
+            IconButton(onClick = {
+                scope.launch {
+                    drawerState.open()
+                }
+            }) {
                 Icon(
                     Icons.Outlined.Menu,
                     contentDescription = stringResource(R.string.home_menu),
@@ -197,8 +624,11 @@ fun HomeHeader(
 @Preview
 @Composable
 fun HomeHeaderPreview() {
+    val scope = rememberCoroutineScope()
+    val drawerState = rememberDrawerState(DrawerValue.Closed)
     HomeHeader(
-        openDrawer = {},
+        scope,
+        drawerState,
         onClickSortType = {},
         onClickListingType = {},
         onClickRefresh = {},
