+++ app/src/main/java/com/jerboa/ui/components/person/PersonProfileActivity.kt
@@ -101,7 +101,6 @@ fun PersonProfileActivity(
     useCustomTabs: Boolean,
     usePrivateTabs: Boolean,
     blurNSFW: Boolean,
-    drawerState: DrawerState,
 ) {
     Log.d("jerboa", "got to person activity")
 
@@ -132,12 +131,6 @@ fun PersonProfileActivity(
                 }
                 else -> {}
             }
-        }
-    }
-
-    fun openDrawer() {
-        scope.launch {
-            drawerState.open()
         }
     }
 
@@ -174,7 +167,7 @@ fun PersonProfileActivity(
                         onClickSortType = {},
                         onBlockPersonClick = {},
                         onReportPersonClick = {},
-                        openDrawer = ::openDrawer,
+                        navController = navController,
                     )
                 }
                 is ApiState.Holder -> {
@@ -223,7 +216,7 @@ fun PersonProfileActivity(
                                 navController.toPostReport(id = firstPost.post.id)
                             }
                         },
-                        openDrawer = ::openDrawer,
+                        navController = navController,
                     )
                 }
                 else -> {}
