+++ app/src/main/res/values-de/strings.xml
@@ -38,6 +38,7 @@
     <string name="dialogs_top_week">Top der Woche</string>
     <string name="dialogs_top_month">Top des Monats</string>
     <string name="dialogs_top_year">Top des Jahres</string>
+    <string name="dialogs_top_all_time">Top aller Zeiten</string>
     <string name="dialogs_subscribed">Abonniert</string>
     <string name="dialogs_local">Lokal</string>
     <string name="dialogs_all">Alle</string>
@@ -70,6 +71,9 @@
     <string name="home_subscribed">Abonniert</string>
     <string name="home_local">Lokal</string>
     <string name="home_all">Alle</string>
+    <string name="home_saved">Gespeichert</string>
+    <string name="home_profile">Profil</string>
+    <string name="home_inbox">Posteingang</string>
     <string name="home_settings">Einstellungen</string>
     <string name="home_subscriptions">Abonnements</string>
     <string name="home_add_account">Account hinzufügen</string>
@@ -253,6 +257,7 @@
     <string name="sorttype_topall">Top über alles</string>
     <string name="sorttype_mostcomments">Meiste Kommentare</string>
     <string name="sorttype_newcomments">Neue Kommentare</string>
+    <string name="home_communities">Communities</string>
     <string name="bottomBar_search">Zur Suche springen</string>
     <string name="bottomBar_label_search">Suche</string>
     <string name="bottomBar_label_home">Startseite</string>
