+++ app/src/main/res/values-ru/strings.xml
@@ -55,6 +55,7 @@
     <string name="dialogs_new_comments">Недавно прокомментированное</string>
     <string name="dialogs_subscribed">Подписки</string>
     <string name="dialogs_top">Лучшее</string>
+    <string name="dialogs_top_all_time">Лучшее за все время</string>
     <string name="dialogs_top_day">Лучшее за день</string>
     <string name="dialogs_top_month">Лучшее за месяц</string>
     <string name="dialogs_top_week">Лучшее за неделю</string>
@@ -66,10 +67,14 @@
     <string name="homeHeader_filter">Выберите фильтр постов</string>
     <string name="home_add_account">Добавить аккаунт</string>
     <string name="home_all">Все посты</string>
+    <string name="home_communities">Сообщества</string>
+    <string name="home_inbox">Сообщения</string>
     <string name="home_local">Локальные посты</string>
     <string name="home_menu">Меню</string>
     <string name="home_post_view_mode">Режим отображения постов</string>
+    <string name="home_profile">Профиль</string>
     <string name="home_refresh">Обновить</string>
+    <string name="home_saved">Закладки</string>
     <string name="home_settings">Настройки</string>
     <string name="home_sign_out">Выйти</string>
     <string name="home_site_info">Информация о сайте</string>
